tests: adição de testes para servicodecomposicao

+++ test/Divisores.Dominio.Tests/NumeroTest.cs
@@ -1,4 +1,4 @@
-using Divisore.Dominio;
+using Divisores.Dominio;
 using FluentAssertions;
 using System;
 using System.Collections.Generic;
@@ -15,7 +15,7 @@ namespace Divisores.Dominio.Tests
 
             var divisores = numero.Divisores();
 
-            divisores.Should().Equal(new HashSet<int> { 1, 3, 5, 9, 15, 45 });
+            divisores.Should().Equal(new HashSet<long> { 1, 3, 5, 9, 15, 45 });
         }
 
         [Fact(DisplayName = "Deve retornar divisores de dois.")]
@@ -25,7 +25,7 @@ namespace Divisores.Dominio.Tests
 
             var divisores = numero.Divisores();
 
-            divisores.Should().Equal(new HashSet<int> { 1, 2 });
+            divisores.Should().Equal(new HashSet<long> { 1, 2 });
         }
 
         [Fact(DisplayName = "Deve retornar divisores de um.")]
@@ -35,7 +35,7 @@ namespace Divisores.Dominio.Tests
 
             var divisores = numero.Divisores();
 
-            divisores.Should().Equal(new HashSet<int> { 1 });
+            divisores.Should().Equal(new HashSet<long> { 1 });
         }
 
         [Fact(DisplayName = "Deve levantar exceção quando numero for zero.")]
